% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bic_weight.R
\name{calc_bic_weights}
\alias{calc_bic_weights}
\title{Calculate BIC weights}
\usage{
calc_bic_weights(bics)
}
\arguments{
\item{bics}{}
}
\value{
A vector with BIC weights
}
\description{
Calculate BIC weights
}
\examples{
calc_bic_weights(c(21, 20, 15, 24))
}
