% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_archipelagos.R
\name{facet_archipelagos}
\alias{facet_archipelagos}
\title{Rates and area archipelago faceted}
\usage{
facet_archipelagos(base_rates, standardisation = FALSE)
}
\arguments{
\item{base_rates}{}
}
\value{
A ggplot with each estimated rate profile and area per archipelago
in a facet
}
\description{
Rates and area archipelago faceted
}
