% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_base_rates.R
\name{get_base_rates}
\alias{get_base_rates}
\title{Calculate base rates per archipelago and model}
\usage{
get_base_rates(archipelago_data, pars, model)
}
\arguments{
\item{model}{}
}
\description{
Calculate base rates per archipelago and model
}
