% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/islandpaleoarea-package.R
\docType{package}
\name{islandpaleoarea-package}
\alias{islandpaleoarea}
\alias{islandpaleoarea-package}
\title{islandpaleoarea: Paleo-Area Influence on Island Evolutionary Models}
\description{
This package aims at incorporating paleo-area data on diversification rates. To do so, it uses the island biogeography model DAISIE.
}
\author{
\strong{Maintainer}: Pedro Santos Neves \email{p.m.santos.neves@rug.nl} (\href{https://orcid.org/0000-0003-2561-4677}{ORCID})

}
\keyword{internal}
