% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_area_time.R
\name{plot_area_time}
\alias{plot_area_time}
\title{Plot total archipelago area through time}
\usage{
plot_area_time(ordered_results)
}
\arguments{
\item{ordered_results}{}
}
\value{
A ggplot2 line plot with the sum of every archipelago area at all
time points
}
\description{
Plot total archipelago area through time
}
\author{
Pedro Santos Neves
}
