% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_sel_time.R
\name{plot_model_sel_time}
\alias{plot_model_sel_time}
\title{Plot model selection area plot}
\usage{
plot_model_sel_time(ordered_results, best_models = c(17, 19, 4, 18))
}
\arguments{
\item{best_models}{}
}
\description{
Plot model selection area plot
}
