% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_res_area.R
\name{plot_res_area}
\alias{plot_res_area}
\title{Plot archipelago area and parameter estimates through time}
\usage{
plot_res_area(ordered_results)
}
\arguments{
\item{ordered_results}{Data frame with all the results, BIC and BIC weights ordered
by age and model for all archipelagos}
}
\value{
A faceted ggplot with the areas and parameters estimates through time
}
\description{
Plot archipelago area and parameter estimates through time
}
