% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_results}
\alias{read_results}
\title{Read model fitting results}
\usage{
read_results(results_path = "results")
}
\arguments{
\item{results_path}{A string with valid path for folder containing model
fitting results as obtained from runs on Peregrine HPCC via the contained
R and bash scripts on this package.}
}
\value{
A data frame with model parameters, model ID, seed and age in ky.
}
\description{
Read model fitting results
}
\author{
Pedro Santos Neves
}
