% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_paper_plot.R
\name{save_paper_plot}
\alias{save_paper_plot}
\title{Save a plot to current directory}
\usage{
save_paper_plot(plot_to_save, file_name, type_size = "full_size")
}
\value{
NULL. A pdf and png files with the plot to save are saved to the
working directory
}
\description{
Save a plot to current directory
}
\examples{
\dontrun{
plotting_table <- prepare_results_to_plot(readRDS("full_res.rds"))
plot <- plot_estimate_results(
  plotting_table,
  parameter_name = "lambda_c",
  partition_by = "island_age",
  colour_by = "stac",
  shape_by = "c_m"
)
save_paper_plot(plot_to_save = plot, "lac_plot", type_size = "full_size")
}
}
\author{
Pedro Santos Neves
}
