% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_mw_model_fixed_pars.R
\name{setup_mw_model_fixed_pars}
\alias{setup_mw_model_fixed_pars}
\title{Setup parameters needed for MW version of DAISIE}
\usage{
setup_mw_model_fixed_pars(model, best_previous_time_slice)
}
\arguments{
\item{model}{A numeric from 1 to 28, corresponding to the model to be run.}
}
\value{
A list with the necessary parameters for the \code{\link[DAISIE:DAISIE_MW_ML]{DAISIE::DAISIE_MW_ML()}} run.
}
\description{
Setup parameters needed for MW version of DAISIE
}
\examples{
model_parameters <- setup_mw_model(model = 1)
}
\author{
Luis Valente, Pedro Santos Neves
}
