% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_pars_model.R
\name{translate_pars_model}
\alias{translate_pars_model}
\title{Title}
\usage{
translate_pars_model(
  area,
  distance,
  Age,
  lamc_0,
  y,
  mu_0,
  x,
  K_0,
  z,
  gam_0timesM,
  alpha,
  lama_0,
  beta
)
}
\arguments{
\item{beta}{}
}
\description{
Title
}
