% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{Default parameter documentation to inherit from}
\usage{
default_params_doc(results_path)
}
\arguments{
\item{results_path}{A string with valid path for folder containing model
fitting results as obtained from runs on Peregrine HPCC via the contained
R and bash scripts on this package.}
}
\value{
Nothing
}
\description{
Default parameter documentation to inherit from
}
\author{
Pedro Santos Neves
}
\keyword{internal}
