% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_sel_time.R
\name{plot_model_sel_time}
\alias{plot_model_sel_time}
\title{Plot model selection area plot}
\usage{
plot_model_sel_time(ordered_results, best_models = c(17, 19, 18))
}
\arguments{
\item{ordered_results}{Data frame with all the results, BIC and BIC weights ordered
by age and model for all archipelagos}

\item{best_models}{A numeric vector with the numbers of the parameters to be plotted}
}
\value{
An area plot. Area corresponds to BIC weights of the models defined
in the \code{ordered_results} vector.
}
\description{
Plot model selection area plot
}
\examples{
data(ordered_results_paleo)
model_sel_plot <- plot_model_sel_time(ordered_results_paleo)
}
